requirements.txt
flask
google-generativeai
python-whois
python-dotenv
